Removing client-side from serverside page
